add works properly now
