+++ src/components/BirdNerd.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import NavBar from "./nav/NavBar"
 import ApplicationViews from "./ApplicationViews"
 import './BirdNerd.css';
 
@@ -7,7 +6,6 @@ export default class BirdNerd extends Component {
   render() {
     return (
       <React.Fragment>
-        <NavBar />
         <ApplicationViews />
       </React.Fragment>
     );
