+++ src/components/nav/NavBar.js
@@ -10,24 +10,9 @@ import {
   NavLink
 } from "reactstrap";
 import SightingModal from "../sighting/SightingModal";
-import APIManager from "../../modules/APIManager";
 import "./NavBar.css";
 
 export default class NavBar extends React.Component {
-  state = {
-    birds: [],
-    sightings: [],
-    date: "",
-    location: "",
-    birdId: "",
-    summary: ""
-  };
-
-  componentDidMount() {
-    APIManager.getAllEntries("birds").then(birds => {
-      this.setState({ birds: birds });
-    });
-  }
 
   constructor(props) {
     super(props);
@@ -44,41 +29,6 @@ export default class NavBar extends React.Component {
     });
   }
 
-  addSighting = sighting => {
-    return APIManager.addEntry("sightings", sighting)
-      .then(() => APIManager.getAllEntries("sightings"))
-      .then(sightings =>
-        this.setState({
-          sightings: sightings
-        })
-      );
-  };
-
-  handleFieldChange = evt => {
-    const stateToChange = {};
-    stateToChange[evt.target.id] = evt.target.value;
-    this.setState(stateToChange);
-  };
-
-  constructNewSighting = () => {
-    if (this.state.bird === "") {
-      window.alert("Please select a bird");
-    } else {
-      const sighting = {
-        date: this.state.date,
-        location: this.state.location,
-        birdId: this.state.birds.find(b => b.name === this.state.birdId).id,
-        summary: this.state.summary
-      };
-      console.log("sighting", sighting);
-      this.addSighting(sighting)
-
-      // FIXME:Make the page reload with state
-      window.location.href = "http://localhost:3000/"
-
-    }
-  };
-
   render() {
     return (
       <React.Fragment>
@@ -92,10 +42,10 @@ export default class NavBar extends React.Component {
               <Nav className="ml-auto" navbar>
                 <NavItem>
                   <SightingModal
-                    handleFieldChange={this.handleFieldChange}
-                    constructNewSighting={this.constructNewSighting}
-                    birds={this.state.birds}
-                    sightings={this.state.sightings}
+                    birds={this.props.birds}
+                    sightings={this.props.sightings}
+                    handleFieldChange={this.props.handleFieldChange}
+                    constructNewSighting={this.props.constructNewSighting}
                   />
                 </NavItem>
                 <NavItem>
