+++ src/components/sighting/SightingCard.js
@@ -15,3 +15,5 @@ export default class SightingCard extends Component {
     );
   }
 }
+
+
