+++ src/components/sighting/SightingList.js
@@ -5,21 +5,19 @@ import SightingCard from "./SightingCard";
 export default class SightingList extends Component {
   state = {
     users: [],
-    sightings: [],
-    birds: [],
     birdName: ""
   };
 
-  componentDidMount() {
-    const newState = {};
+  // componentDidMount() {
+  //   const newState = {};
 
-    APIManager.getAllEntries("sightings", "?_sort=date&_order=asc&_expand=bird")
-      .then(sightings => {
-        this.setState({ sightings: sightings });
-      })
+  //   APIManager.getAllEntries("sightings", "?_sort=date&_order=asc&_expand=bird")
+  //     .then(sightings => {
+  //       this.setState({ sightings: sightings });
+  //     })
 
-      .then(() => this.setState(newState));
-  }
+  //     .then(() => this.setState(newState));
+  // }
 
   render() {
     return (
@@ -27,11 +25,13 @@ export default class SightingList extends Component {
         <h2>Sightings</h2>
         <section className="sightings">
           <div className="card__holder">
-            {this.state.sightings.map(sighting => (
+            {this.props.sightings.map(sighting => (
               <SightingCard
                 key={sighting.id}
                 sighting={sighting}
-                sightings={this.state.sightings}
+                sightings={this.props.sightings}
+                birds={this.props.birds}
+                // birdName={this.state.birdName}
               />
             ))}
           </div>
